refactor(generator): extract build_column_tests and normalize_test_status to lookups
Move _build_column_tests and normalize_test_status out of transforms/models.py
into transforms/lookups.py so transform_source can call them too. Generalize
the parameter name to test_nodes_by_id since the map is keyed on any
unique_id, not only models.

Re-export normalize_test_status from models.py to preserve the existing
import path used by generator/erd.py and tests.

No behavior change.

Prep for DOC-226.
